==> Component_Demos/00B_Package_Descr_Refs.Rmd <==
---
title: "Package descriptions and helpful links"
author: "Matt Multach"
date: "`r Sys.Date()`"
output: 
  bookdown::html_document2:
    toc: true
    toc_depth: 4
    toc_float: 
       collapsed: true
       smooth_scroll: true
    number_sections: true
    #pandoc_args: --top-level-division=section
bibliography: ../References/Refs_Main.bib
---

# File Introduction, Preamble and Setup {.unnumbered .tabset .tabset-fade .tabset-pills}

```{r setup, message = FALSE , warning = FALSE , include=FALSE}
knitr::opts_chunk$set(echo = TRUE , eval = TRUE)
```

<style>
    .ref {
        position: relative;
        vertical-align: baseline;
    }

    .refnum {
        position: relative;
        left: -1px;
        bottom: 1ex;
        font-family: Verdana, sans-serif;
        color: #005994;
        font-size: .7em;
        font-weight: bold;
        text-decoration: underline;
        cursor: pointer;        
    }

    .refbody {
        font-family: Verdana, sans-serif;
        font-size: .7em;
        line-height: 1.1;
        display: none;
        min-width: 20em;
        position: absolute;
        left: 25px;
        bottom: 5px ;
        border: 1px solid;
        padding: 5px;
        background-color: #fff;
        word-wrap: break-word;
        z-index: 9999;
        overflow: auto;
    }

</style>

<script>
    function footypop(id) {
        var el = document.getElementById(id) ;
  
        if (el.style.display == "none") {
            el.style.display = "block" ;
        } else {
            el.style.display = "none" ;
        }
        
    }
</script>

## Introduction {.unnumbered}

This document is intended as a central glossary of the packages used throughout these demonstrations. The main goals include:
 * A description of each package used, including an outline of its utility and how it is used
 * A review of the most useful functions used in the course of the various demos in this repository
 * Links to and suggestions for external resources for learning more about these packages, often including work done by the original package authors and developers.
 
These packages are currently presented in no particular order. Once the initial set of packages are included and described, subsequent revisions will group packages according to type of package or functionality accomplished by the package.

## Formatting Notes {.unnumbered}

This document makes use of **bolding**, _italics_, and "quotations" to help distinguish references to different types of items.

 * **bolding** will be used when referencing a new term or concept for the first time. Subsequent references to each term/concept will _not_ be bolded.
 * _italics_ will be used primarily for emphasis.
 * 'single quotations' will be used to clarify specific arguments for a function, or specific parameters of a mathemtical/statistical formulation
 * Inline references to code (functions, objects, specific commands, etc.) will use **<span style = "color:blue">`` `r "code_chunk_formatting"` ``</span>** in **<span style = "color:blue">`` `r "bolded blue font"` ``</span>**
 * Inline references to packages will similarly use *<span style = "color:red">`` `r "code_chunk_formatting"` ``</span>*, except in *<span style = "color:red">`` `r "italicized red font"` ``</span>*
 * References to other documents in this repository will use <span style = "color:green">`` `r "code_chunk_formatting"` ``</span> in <span style = "color:green">`` `r "un-italicized, un-bolded, green font"` ``</span>
 
The most important aspects of how I use spacing and line breaks are detailed below.

I generally put spaces between code inputs I consider "sufficiently distinct". This improves readability generally, but I find it particularly helpful for debugging. Note, however, that spaces are generally trivial in between distinct code elements in R, although this is not universally the case. Multi-character code elements, such as the pointer **<span style = "color:blue">`` `r "<-"` ``</span>** and most logical operators, _cannot_ include spaces in between components of the element. Note also that whitespace *is* meaningful in other programming languages, and so this convention should be considered with caution in your own practice.

Generally, I use line breaks to:

 * Break up separate arguments for a single command/function or chain of operations
 * To clearly distinguish between different closing parentheses, brackets, squigglies, etc., since RStudio will automatically tab each closing piece to align with its opening segment.


# Data Manipulation and Programmatic Tools {.tabset .tabset-fade .tabset-pills}

## *<span style = "color:red">`` `r "magrittr"` ``</span>*, @magrittr

### Description {.unnumbered .unlisted}

The *<span style = "color:red">`` `r "magrittr"` ``</span>* package implements the pipe operator <span style = "color:blue">`` `r "%>%"` ``</span> in R to facilitate operations on objects. 

### Functions Used {.unnumbered .unlisted}

Aside from the pipe operator <span style = "color:blue">`` `r "%>%"` ``</span>, no other functions from *<span style = "color:red">`` `r "magrittr"` ``</span>* were utilized in these demos.  

### External Resources {.unnumbered .unlisted}

*<span style = "color:red">`` `r "magrittr"` ``</span>* is a part of the [`r "tidyverse"`](https://www.tidyverse.org/) and therefore many guides and walkthroughs have been create to explore its usage. The `r "tidyverse"` website has a [section devoted to the package](https://magrittr.tidyverse.org/). Hadley Wickham, a major contributor to and proponent of the `r "tidyverse"`, also has two books on using R that each discuss *<span style = "color:red">`` `r "magrittr"` ``</span>* and piping to varying degrees. [Chapter 18](https://r4ds.had.co.nz/pipes.html?q=magrittr#prerequisites-11) of R for Data Science<span class="ref"><span class="refnum" onclick="footypop('ref1')">[ref]</span><span id="ref1" class="refbody">@R4DS</span></span> provides a comprehensive, ground-level review of implementing the pipe operator. More advanced R users can refer to [Chapter 6](https://adv-r.hadley.nz/functions.html#function-composition) of Wickham's Advanced R Book<span class="ref"><span class="refnum" onclick="footypop('ref2')">[ref]</span><span id="ref2" class="refbody">@AdvancedR</span></span>.

Both books are freely available online, and I strongly encourage checking them out!



## *<span style = "color:red">`` `r "purrr"` ``</span>*, @purrr

### Description {.unnumbered .unlisted}

*<span style = "color:red">`` `r "purrr"` ``</span>*, like *<span style = "color:red">`` `r "magrittr"` ``</span>* previously, is a member of the `r "tidyverse"`. The package implements additional tools for functional programming. I use it primarily for its additional features for mapping functions over object spaces.

### Functions Used {.unnumbered .unlisted}

A variety of **<span style = "color:blue">`` `r "map()"` ``</span>**-related functions are used throughout my demos to efficiently iterate and apply functions over a collection of objects or multiple components within a single complicated object.

Additionally, the **<span style = "color:blue">`` `r "safely()"` ``</span>** was one of the most useful functions I discovered over the course of my PhD research. This allows you to run code without fear of the function outright stopping in the face of an error. When running hundreds of thousands (not joking) of models for my primary research, I really did **not** want my progress to halt because one round ran into an error. **<span style = "color:blue">`` `r "safely()"` ``</span>** will run whatever is passed to it and produce 2-element list as a result. One list element will produce the results of a completed function, while the second list element will populate only if the function being passed runs into an error. In that case, the corresponding error message will be stored to this list element. This allows a mapped function to continue running when it hits an error on an iteration rather than completely stopping its progress. 

### External Resources {.unnumbered .unlisted}

[Tidyverse website section](https://purrr.tidyverse.org/)

The [Iteration chapter](https://r4ds.had.co.nz/iteration.html) of _R for Data Science_ heavily features this package.

The [Functionals chapter](https://adv-r.hadley.nz/functionals.html) of _Advanced R_ likewise focuses heavily on related concepts.







## *<span style = "color:red">`` `r "data.table"` ``</span>*, @data.table

### Description {.unnumbered .unlisted}

Implements the <span style = "color:blue">`` `r "data.table"` ``</span> object and includes tools for managing, manipulating, and generally using them. Comparable to Python's *<span style = "color:red">`` `r "numpy"` ``</span>* library.

### Functions Used {.unnumbered .unlisted}

The package's utility in these demos comes primarily from operations on data.tables.

### External Resources {.unnumbered .unlisted}

The package's [main web page](https://rdatatable.gitlab.io/data.table/), which includes a Cheat Sheet.

[This blog](https://www.machinelearningplus.com/data-manipulation/datatable-in-r-complete-guide/) was very helpful to me as I first started learning how to use *<span style = "color:red">`` `r "data.table"` ``</span>*.



## *<span style = "color:red">`` `r "rlang"` ``</span>*, @rlang

### Description {.unnumbered .unlisted}

A collection of tools for interacting with R's core language characteristics.

### Functions Used {.unnumbered .unlisted}

I only used one function from this package: **<span style = "color:blue">`` `r "parse_expr()"` ``</span>**, which turns a character string into a code-usable expression. 

### External Resources {.unnumbered .unlisted}

The package's [main website](https://rlang.r-lib.org/), which includes a Cheat Sheet.

## General Resources

[MSDR](https://bookdown.org/rdpeng/RProgDA/)<span class="ref"><span class="refnum" onclick="footypop('ref3')">[ref]</span><span id="ref3" class="refbody">@MSDR</span></span> and [Advanced R](https://adv-r.hadley.nz/index.html)<span class="ref"><span class="refnum" onclick="footypop('ref4')">[ref]</span><span id="ref4" class="refbody">@AdvancedR</span></span> each cover a variety of advanced topics in developing with and programming in R



# Visualization {.tabset .tabset-fade .tabset-pills}

[This chapter](https://bookdown.org/rdpeng/RProgDA/building-data-visualization-tools.html) provides an extensive overview of more advanced topics in visualization in R<span class="ref"><span class="refnum" onclick="footypop('ref5')">[ref]</span><span id="ref5" class="refbody">@MSDR</span></span>.


## *<span style = "color:red">`` `r "ggplot2"` ``</span>*, @ggplot2

### Description {.unnumbered .unlisted}

Another `r "tidyverse"` package that provides an extensive set of tools for visualization.

### Functions Used {.unnumbered .unlisted}

So many, and many aesthetic mappings within functions, aka "layers" on top of created visual objects.

### External Resources {.unnumbered .unlisted}

[Tidyverse website](https://ggplot2.tidyverse.org/), plus Cheat Sheet

The [Visualization chapter](https://r4ds.had.co.nz/data-visualisation.html) of _R for Data Science_ focuses on this package.



## *<span style = "color:red">`` `r "gridExtra"` ``</span>*, @gridExtra

### Description {.unnumbered .unlisted}

To be honest, I only used this package to be able to create a text-based grob (grid object) for titling my multi-plot visualization.

### Functions Used {.unnumbered .unlisted}

**<span style = "color:blue">`` `r "textGrob()"` ``</span>** makes a grid object out of a character string.

### External Resources {.unnumbered .unlisted}

The visualation tools chapter in MSDR<span class="ref"><span class="refnum" onclick="footypop('ref6')">[ref]</span><span id="ref6" class="refbody">@MSDR</span></span> features [a section](https://bookdown.org/rdpeng/RProgDA/the-grid-package.html) about using the package for developing your own visual objects.




# Document & Presentation Creation in R {.tabset .tabset-fade .tabset-pills}

## *<span style = "color:red">`` `r "xaringan"` ``</span>*, @xaringan

### Description {.unnumbered .unlisted}

I _cannot_ extoll the virtues of the *<span style = "color:red">`` `r "xaringan"` ``</span>* package enough. The author, Yihui Xie, will come up again in this document, as he is one of the biggest contributors to document and presentation functionality in RStudio. 

I don't actively use this package all that often, but the passive support it provides while putting together Markdown reports incredible. The main benefit of this package is to allow for an actively-updating view of a compiled markdown document within RStudio itself in the View pane. The package also adds functionality for making presentations, being based heavily on remark.js.

### Functions Used {.unnumbered .unlisted}

No specific functions are used from this package throughout these various demos.

### External Resources {.unnumbered .unlisted}

[Yihui Xie's extensive R Markdown ebook](https://bookdown.org/yihui/rmarkdown/xaringan.html) is absolutely free and includes an entire chapter on using *<span style = "color:red"> `` `r "xaringan"` ``</span>*. The link leads directly to that chapter.

[Yihui also created a slide presentation using *<span style = "color:red"> `` `r "xaringan"` ``</span>*](https://slides.yihui.org/xaringan/#1)

### Nerd Things {.unnumbered .unlisted}

This next section is for individuals who immediately thought *<span style = "color:red"> `` `r "xaringan"` ``</span>* looked familiar except for _one_ letter. The answers to this tickling sense of familiarity can be found in the last few slides of the slideshow presentation, where Yihui explains the naming conventions he used while developing this package. 

As it turns out, X in some Chinese dialects is pronounced "sh", so the name of the package is actually read as "sharingan," aka that thing you thought of when you first read it. Reading into some of the package functionality, the connection with the manga/anime _Naruto_ becomes even more explicit. For example: the tool within *<span style = "color:red"> `` `r "xaringan" ` ``</span>* that renders the actively-updating markdown document within RStudio is named "Infinite Moon Reader," aka the Infinite Tsukuyomi.

I hope a few people get as much joy out of this revelation as I did!





# Packages for Implementing Specific Statistical Models {.tabset .tabset-fade .tabset-pills}

I will not spend too much time explaining the models implemented by each of the following packages. Please see <a href="https://multach87.github.io/Robust_Lasso_ElasticNet/Component_Demos/00A_Lasso_ENet_Adaptations.html"><span style = "color:green">`` `r "00A_Lasso_ENet_Adaptations"` ``</span></a> for an introduction the lasso, the elastic net, and the relevant adaptations, including rationale, mathematical formulation, and computational implementation.

## *<span style = "color:red">`` `r "glmnet"` ``</span>*, @glmnet

### Description {.unnumbered .unlisted}

A package for fitting lasso- and elastic net-regularized regression models with a variety of link functions.

### Functions Used {.unnumbered .unlisted}

This package is mostly included because the implementation for the lasso/elastic net adaptations of interest all rely on it. Additionally, I make use of the **<span style = "color:blue">`` `r "cv.glmnet()"` ``</span>** for selecting the optimatl value for the adaptive lasso tuning hyperparameter. This function conducts a cross-validation-driven regularization model.

### External Resources {.unnumbered .unlisted}

[The main website](https://glmnet.stanford.edu/) for the package.



## *<span style = "color:red">`` `r "hqreg"` ``</span>*, @hqreg

### Description {.unnumbered .unlisted}

This package adds functionality on top of the *<span style = "color:red">`` `r "glmnet"` ``</span>* model functions to incorporate additional loss functions, including quantile/LAD loss and Huber loss.

### Functions Used {.unnumbered .unlisted}

**<span style = "color:blue">`` `r "cv.hqreg()"` ``</span>** implements cross-validated lasso/ridge/elastic net regression but allows the use of quantile and Huber loss functions. 

### External Resources {.unnumbered .unlisted}

See the [CRAN package guide](https://cran.r-project.org/web/packages/hqreg/hqreg.pdf) as well as the [related paper](https://arxiv.org/abs/1509.02957) for more details.



## *<span style = "color:red">`` `r "msaenet"` ``</span>*, @msaenet

### Description {.unnumbered .unlisted}

Implements a multiple-step adaptive procedure for the lasso and elastic net.

### Functions Used {.unnumbered .unlisted}

**<span style = "color:blue">`` `r "msaenet()"` ``</span>** implements the multi-step adaptive elastic net.

### External Resources {.unnumbered .unlisted}

[The main website](https://nanx.me/msaenet/) for the package.

[The related paper](https://www.tandfonline.com/doi/abs/10.1080/00949655.2015.1016944?journalCode=gscs20).





# Mathematical Functionality and Other Statistical Tools {.tabset .tabset-fade .tabset-pills}

## *<span style = "color:red"> `` `r "mvtnorm" ` `` </span>*, @mvtnorm

### Description {.unnumbered .unlisted}

This package deals with multivariate-normal data and relevant distributional tests and characteristics.

### Functions Used {.unnumbered .unlisted}

My data-generating function, **<span style = "color:blue">`` `r "data_gen()" ` ``</span>**, uses the **<span style = "color:blue">`` `r "rmvnorm"` ``</span>**. I applied the function to generate data values from multivariate normal distributions with one of two characteristic sets (uncontaminated or outlier-contaminated). Three arguments were used to accomplish this:

 * 'n' = number of values to be generated
 * 'mean' = the mean of each of 'p' variables whose values will be generated
 * 'sigma' = the covariance matrix used to generate the values

### External Resources {.unnumbered .unlisted}

The [CRAN reference manual](https://cran.r-project.org/web/packages/mvtnorm/mvtnorm.pdf) describes the various functions in the package and their usage.



## *<span style = "color:red">`` `r "WRS2"` ``</span>*, @WRS2

### Description {.unnumbered .unlisted}

A collection of functions for implementing robust hypothesis-testing techniques and procedures.

### Functions Used {.unnumbered .unlisted}

An adapted version of **<span style = "color:blue">`` `r "gandh()"` ``</span>** for generating values from the **g-and-h distribution**. More details on this distribution can be found in <a href="https://multach87.github.io/Robust_Lasso_ElasticNet/Component_Demos/01_Data_Generation.html"><span style = "color:green">`` `r "01_Data_Generation"` ``</span></a>.

### External Resources {.unnumbered .unlisted}

A publication-length [series of vignettes](https://cran.r-project.org/web/packages/WRS2/vignettes/WRS2.pdf) introducing the package.





# References {.unnumbered}
